Changed difference calculation to use doubles for TFIDF, and propagate TFIDF calculation into difference calculation and output

+++ src/main/java/edu/virginia/aid/comparison/GenericDifference.java
@@ -4,7 +4,7 @@ public class GenericDifference extends Difference {
 
     private String differenceMessage;
 
-    public GenericDifference(String differenceMessage, int differenceScore) {
+    public GenericDifference(String differenceMessage, double differenceScore) {
         super(differenceScore);
 
         this.differenceMessage = differenceMessage;
